* lisp/font-lock.el: No longer use headings as end of section markers.
Each section ends right before the following section begins and IMO
that means that it is unnecessary to mark the end of sections.

For users of `outline-minor-mode' the old end-of-section markers were
a distraction.  They made it much harder to parse the overview outline
state because each section heading was followed by a end-of-section
marker that was formatted as a section heading.  Because of this I
wanted to remove the end-of-section markers.

But as Eli pointed out these sections are long and not everyone uses
`outline-minor-mode'.

So instead of removing them, I am turning the end-of-section markers
into regular comments (beginning with just two semicolons) instead of
section headings (beginning with tree semicolons).  That way users of
`outline-minor-mode' won't be distracted by them and others can still
benefit from the markers as before.