Enable replying to an ical event even when not an attendee
* lisp/gnus/gnus-icalendar.el
(gnus-icalendar-event--build-reply-event-body): Display a warning
instead of barfing when user is missing from attendee list.

When the user identity is not present in the attendee list, an error
is triggered making replying to such an event impossible (the reply
message not being even composed).  This replaces it with a warning.
This is necessary because one may receive events that the organizer
did not set up well and it is up to the user to decide whether or not
to reply to them (bug#41723).
Copyright-paperwork-exempt: yes