Simplify and rely more in the post-command-hook
* lisp/completions-highlight.el (completions-highlight-this-completion)
: Use the new function completions-highlight-select-near.
(completions-highlight-next-completion) : Removed
(completions-highlight-select-near) : New function to select and
highlight next candidate closer to point.
(completions-highlight-previous-completion) :
(completions-highlight-next-line-completion) :
(completions-highlight-previous-line-completion) : Removed as they are
not needed with the hook.
(with-minibuffer-scroll-window) : Add call to
completions-highlight-this-completion at the end.
(minibuffer-next-completion) :
(completions-highlight-previous-completion) :
(completions-highlight-next-line-completion) :
(minibuffer-previous-line-completion) : Remove reference to removed
functions.
(completions-highlight-completions-map) : Clear the bindings for the
removed functions.
(completions-highlight-setup) : Set
completions-highlight-this-completion in post-command-hook instead of
isearch-mode-end-hook.