Support nil value for dictionary-server
* net/lisp/dictionary.el (dictionary-server): Support choice to select
the dictionary server to use
* net/lisp/dictionary.el (dictionary-check-connection): Support nil
value for dictionary-server

This nil value is the new default value of that variable. When
opening a new connection and dictionary-server is nil the code
behaves the following way:

- it will first try to connect to a dictd server running on localhost
- if that fails, it queries the user if the alternative
  server (dict.org) should be consulted
- if the user agrees, the connection is made to dict.org

This allows the default value of dictionary-server not to connect
a remote server by default. The user is always able to select a
different server by customizing the variable dictionary-search.